backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.5
argon2-cffi==23.1.0

# Async Task Processing
celery==5.3.4
flower==2.0.1

# API Integration
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.1

# Validation & Serialization
pydantic==2.5.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.32.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# AWS Integration
boto3==1.29.3
botocore==1.32.3

# PDF Generation (for reports)
reportlab==4.0.7
weasyprint==60.1

# Excel Processing (for reports)
openpyxl==3.1.2
xlsxwriter==3.1.9

# Image Processing (for KYC)
Pillow==10.1.0
python-magic==0.4.27 